refactored router / nav into its own component
